requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0.0
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=4.5.0
python-decouple>=3.8
requests>=2.31.0
django-extensions>=3.2.0
pillow>=10.0.0